asset_scrapper/requirements.txt
atomicwrites==1.4.1
attrs==21.4.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
dnspython==2.2.1
Flask==2.1.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.8.0
MarkupSafe==2.1.1
multitasking==0.0.10
numpy==1.22.4
packaging==21.3
pandas==1.4.2
pluggy==1.0.0
py==1.11.0
pymongo==4.2.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.9
Werkzeug==2.1.2
yfinance==0.1.70
